Update source files

+++ source/block.cpp
@@ -50,6 +50,6 @@ void Block::UndoRotation()
     rotationState--;
     if (rotationState == -1)
     {
-        rotationState = cells.size() - 1;
+        rotationState = static_cast<int>(cells.size()) - 1;
     }
 }
